projects/01_fyyur/starter_code/requirements.txt
babel==2.9.0
python-dateutil==2.6.0
Flask-Moment==1.0.2
Flask-WTF==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
psycopg2==2.9.3
